PythonCodeCompletion: 1. implement goto declaration, 2. back to using piped XML with a new fix that should hopefully work with 64 bit windows

+++ PythonCodeCompletion/xmlrpc_embedder.cpp
@@ -1,3 +1,5 @@
+#include "stdint.h"
+
 #include "xmlrpc_embedder.h"
 #include <wx/listimpl.cpp>
 #include <wx/arrimpl.cpp>
@@ -137,14 +139,14 @@ public:
             std::cout<<"bad request value"<<std::endl;
             return false;
         }
-        unsigned long r_size=msg.size();
-        m_ostream->Write(&r_size,sizeof(unsigned long));
+        uint32_t r_size=msg.size();
+        m_ostream->Write(&r_size,sizeof(uint32_t));
         m_ostream->Write(msg.c_str(),msg.size());
         std::cout<<"wrote "<<msg<<std::endl;
         std::cout<<"wrote "<<r_size<<std::endl;
         char ch=m_istream->GetC();
         std::cout<<"read ch "<<ch<<std::endl;
-        for(int i=0;i<sizeof(unsigned long);i++)
+        for(int i=0;i<sizeof(uint32_t);i++)
             ((char*)(&r_size))[i]=m_istream->GetC();
         std::cout<<"response size is "<<r_size<<std::endl;
         char *buf = new char[r_size+1];
